LPS-90226 Move every class that extends or implements a jaxrs class/interface to a jaxrs package. Classes that use jaxrs annotations are not in the jaxrs package.
