+++ modules/apps/portal-vulcan/portal-vulcan-impl/src/main/java/com/liferay/portal/vulcan/internal/jaxrs/context/provider/PaginationContextProvider.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.vulcan.internal.context.provider;
+package com.liferay.portal.vulcan.internal.jaxrs.context.provider;
 
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.vulcan.context.AcceptLanguage;
-import com.liferay.portal.vulcan.internal.context.AcceptLanguageImpl;
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.vulcan.context.Pagination;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -26,24 +25,21 @@ import org.apache.cxf.jaxrs.ext.ContextProvider;
 import org.apache.cxf.message.Message;
 
 /**
- * @author Cristina González
+ * @author Zoltán Takács
  */
 @Provider
-public class AcceptLanguageContextProvider
-	implements ContextProvider<AcceptLanguage> {
-
-	public AcceptLanguageContextProvider(Portal portal) {
-		_portal = portal;
-	}
+public class PaginationContextProvider implements ContextProvider<Pagination> {
 
 	@Override
-	public AcceptLanguage createContext(Message message) {
+	public Pagination createContext(Message message) {
 		HttpServletRequest httpServletRequest =
 			ContextProviderUtil.getHttpServletRequest(message);
 
-		return new AcceptLanguageImpl(httpServletRequest, _portal);
-	}
+		int itemsPerPage = ParamUtil.getInteger(
+			httpServletRequest, "pageSize", 20);
+		int pageNumber = ParamUtil.getInteger(httpServletRequest, "page", 1);
 
-	private final Portal _portal;
+		return Pagination.of(itemsPerPage, pageNumber);
+	}
 
 }
